Improve system call support
This change rewrites a lot of the system calls that use files. We map
magic numbers better. This change breaks signal handling temporarily,
but it's going to get better than ever before soon.

New profiling data is now included in blink/machine.c which shows our
test coverage and how frequently used each dispatch path is so we can
know where the hotspots are in this virtual machine.