Introduce abstraction for JIT function threading
The JIT module is able to generate native functions in memory at
runtime, which make function calls to other functions. We're not
dispatching jitted code paths yet, but we will shortly. The impl
supports x86-64 and aarch64 but it'd be nice to support more.

This change also makes crash reports friendlier, with backtraces
and register information.

I2022-11-20T22:30:41.470104:blink/throw.c:69: 30886: UNDEFINED INSTRUCTION
	 PC 400078 ud2
	 AX 0000000000000000  CX 0000000000000000  DX 0000000000000000  BX 0000000000000000
	 SP 00007ffffffffbf0  BP 0000000000000000  SI 0000000000000000  DI 0000000000000000
	 R8 0000000000000000  R9 0000000000000000 R10 0000000000000000 R11 0000000000000000
	R12 0000000000000000 R13 0000000000000000 R14 0000000000000000 R15 0000000000000000
	 FS 0000000000000000  GS 0000000000000000 OPS 1                JIT 0
	o/dbg/x86_64/test/asm/add.elf
	000000000000 000000400078 _start

Eventually when the virtual machine is completely stable, those will
just be SIGILL signals sent to the program, similar to how we send a
SIGFPE instruction on divide errors.

MicroBlaze support is dropped because the static qemu binary for it
keeps crashing.