Perform some refactorings
This change makes the code simpler and more elegant by declaring two
macros (P and A) which define the dispatch parameters and arguments.
They're written in the source code so often that it doesn't make sense
to have a big descriptive name. This change also removes evil typedefs.
There's also a few other sneak inconsequential changes.