Handle signal interrupts better
This change gets the https://ipv4.games/ web server working under Blink,
which is multi-threaded and relies on signals for cancellation. Syscalls
that raise EINTR will now spin if the sig wasn't delivered to the guest.
Blink now uses SIGSYS as its method of interrupting i/o calls in threads
which should be reasonable and portable, since we don't emulate SECCOMP.

This change also implements sched_setaffinity() and sched_getaffinity(),
which are basically no-ops on non-Linux, although the latter function is
able to serve the use case of obtaining the CPU count.

Integration tests have been added for testing multi-threaded SQLite WAL,
as well as other tests blocked earlier by the lack of interrupt support.