pthread hack for Haiku OS
Haiku's pthread has a mechanism that compares the current thread ID
with the "owning" thread ID before attempting to unlock a pthread_mutex.

This is problematic in the event of a `fork`: The stored owner thread ID
does not change while the thread ID in the child has changed. This keeps
the mutex locked forever.

This patch currently bypasses this check by setting the owner thread ID
to the current thread ID.