Fix bug in MOVSW, MOVSD, and MOVSQ
I made the classic mistake of using memcpy() when I should have used
memmove(). I never noticed this until now, because in Cosmopolitan Libc
I defined memcpy() to be the same thing as memmove(), because there's no
performance loss in doing so. But when Blink is compiled with other Libc
implementations, e.g. Musl, those lines become problematic when the
destination and source arguments overlap.