VFS build support
Build support for the VFS feature. This feature is disabled by default
and can be enabled by passing --enable-vfs to configure.

VFS and overlays are two mutually exclusive features.

At this stage, --enable-vfs does not work yet and the build will fail
with undefined symbol errors. The default build is unaffected.