Fix CTRL-C bug in TUI under turbo mode
The issue was, if you:

1. Press turbo (`C-t`) a few times
2. Press continue (`c`)
3. Press interrupt (`C-c`)

Then the interface would halt on a stale rendering of the execution
state, which wouldn't be updated until some other keystroke caused
Redraw() to be called. This could be misleading. For example,
interrupting execution to adjust the TURBO setting which look as though
the turbo setting was stepping the program's execution. This change
fixes things by calling ReactiveDraw() at the correct moment.