Fix SHUFPD
This change fixes #68 which was caused by a copy and paste error in the
implementation of the SHUFPD instruction. Regression test is now added.
It's now possible to build the Cosmopolitan repository, e.g.

    cd ~/cosmo
    blink make -j8 o//examples
    blink make -j8 o//tool/net/redbean.com

It's not possible yet to `blink make` everything in cosmo yet, since it
runs tests too, and cosmo tests have platform detection code that'll be
updated soon to accommodate blink.