VFS feature
This commit comes with the full VFS feature. A built-in `hostfs` is also
included. `procfs` and `devfs` are currently not implemented but exist
as an alias to `hostfs` that by default maps to the `/proc` and `/dev`
folders of the host.

Now, all system calls that requires a file descriptor should be
redirected to the corresponding function with a `Vfs` prefix. When this
feature is disabled (the default behavior), these functions will
become macros that redirect to the original ones (or functions
with `Overlays` prefix if they are available).

Two blink syscalls are also fixed:
- `sendfile` now does not cause an infinite loop.
- `select` is now implemented using `poll` and respects the custom
`poll` implementations of each file descriptor.