Fix TUI so it can step into fault signal handlers
In cases where (1) the guest has a signal handler for something like
SIGSEGV and (2) a segmentation fault is generated while in TUI mode,
then Blinkenlights' TUI would switch from Tui() mode to Exec() mode,
which prevented the interface from being able to see and control the
signal delivery handler. This change fixes that, so the TUI will let
the fault signal handler be viewed and debugged.