Fix issues with mprotect()
On systems with a page size larger than 4096, mprotect() was not
functioning properly in certain cases. For example, if virt+size
wasn't aligned to the host page size then host protection wasn't
getting relaxed to read+write which could cause the sub-pages on
the right-hand side to have an invalid host memory protection.

This change also hardens mprotect() against race conditions with
munmap() and mmap() in cases where the guest is behaving weirdly
or intentionally trying to crash a system call.