Support read-only memory protection
Blink previously did very little to enforce the memory protections.
Read-only protections would oftentimes be ignored as read+write, if
convenient. This change ensures that Blink, in all operating modes,
will enforce the guest's memory protections, to the greatest extent
possible. As an exception, if the guest does not accurately observe
the underlying host system page size, then Blink reserves the right
to relax permissions, so that programs will actually work, which is
more important than imposing restrictions on the program's behavior