Fix issues in fast versions of rep movsb & rep stosb
The diactual & siactual pointers are now offset by the ES &
DS bases after the first iteration of the outer loop.  This is
needed for most real-mode uses of rep movsb to work.

I also try to handle the (hopefully rare) case of DI or SI
"wrapping around" to 0 or -1 in the middle of a rep movsb or
rep stosb.