Loosen ELF restrictions
This change loosens the restrictions on EI_OSABI and EI_CLASS so that
Blink is as permissive as the Linux Kernel when it comes to the kinds
of ELF format abuse that are popular in the size coding community. To
make this as safe as possible, ELF overflow checks have been added.