Pass __program_executable_name to cosmo libc
We recently introduced an improved system to Cosmo Libc and Ape Loader
for detecting the path of the executable. This caused issues for blink
because Cosmo was choosing `/proc/self/exe` rather than `argv[0]`. The
solution is to use Cosmo CRT's new feature that lets us pass the value
in the the RDX register to _start(). This allows programs like redbean
which need to introspect their own zip assets, to start working again.

Fixes #162