Use PROT_MPROTECT for NetBSD mprotect restrictions (#185)
NetBSD's mprotect seems to be more restrictive than other OSes, where it
does not allow less restrictive mappings than the original mmap, and can
be subject to PaX restrictions. According to NetBSD documentation, there
is a PROT_MPROTECT macro to define allowed protections for later uses of
mprotect, without granting the permissions immediately in mmap. This can
be used to ensure that the full range of protections blink could make in
the course of execution are permitted.