Enable compilation on Solaris
These changes allow Blink to be built with the default configuration on
Solaris, tested on a Solaris 11.4 VM. The main addition is to introduce
emulation of flock with fcntl, since the former is not available on new
Solaris versions.

Additionally, fd_set handling is modified to use heap memory instead of
stack memory, since this can exceed the frame size.